pyproject.toml
[project]
name = "hanzi-styles-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "pillow>=12.1.0",
]
